customize transporter information accordian button and rename title of transporter row action buttons to better reflect behavior and clean up
