+++ client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "haztrak",
-  "version": "0.7.0",
+  "version": "0.6.2",
   "private": true,
   "scripts": {
     "start": "vite",
