+++ client/src/App.scss
@@ -88,6 +88,10 @@ $info: #009999;
   transform: rotate(-90deg);
 }
 
+.rotate-90 {
+  transform: rotate(90deg);
+}
+
 .app-container {
   min-height: 350px;
 }
