--- client/src/components/Manifest/Transporter/TransporterTable/index.ts
@@ -1,4 +1,0 @@
-import { TransporterTable } from './TransporterTable';
-import { TransporterRowActions } from './TransporterRowActions';
-
-export { TransporterTable, TransporterRowActions };
